+++ src/main/java/org/example/springbootdeveloper/user/domain/User.java
@@ -1,4 +1,4 @@
-package org.example.springbootdeveloper.domain;
+package org.example.springbootdeveloper.user.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
