+++ src/main/java/org/example/springbootdeveloper/user/service/UserService.java
@@ -1,9 +1,9 @@
-package org.example.springbootdeveloper.service;
+package org.example.springbootdeveloper.user.service;
 
 import lombok.RequiredArgsConstructor;
-import org.example.springbootdeveloper.domain.User;
-import org.example.springbootdeveloper.dto.AddUserRequest;
-import org.example.springbootdeveloper.repository.UserRepository;
+import org.example.springbootdeveloper.user.domain.User;
+import org.example.springbootdeveloper.user.dto.AddUserRequest;
+import org.example.springbootdeveloper.user.respository.UserRepository;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
